Create the bridge app using `hono`
Create `createBridgeApp` function which takes an instance of
the Deno KV client and creates a tiny web server app that provides
an API to access the Deno KV Database from frontend web apps or
any other place where where Deno KV client is not usable.
This function returns a web server app that can be used to serve
requests in an HTTP server created by Deno or Node.