Fix "kv.list method throws when it isn't supposed to" issue
Refactor `validateBrowseRequestParams` function and make it
return "prefix", "start", and "end" parameters in one object.
And most inportantly, if one of these three parameters didn't
present, do not include it with the return value as `param: undefined`
because kv.list method of "@deno/kv" package will trate the param
as existing even when the value of the param is undefined and throw
errors in cases when there should not be errors.